Split ColumnUnknownException into table-scoped and unscoped
