+++ server/src/main/java/io/crate/exceptions/unscoped/ColumnUnknownException.java
@@ -6,7 +6,7 @@
  * you may not use this file except in compliance with the License.  You may
  * obtain a copy of the License at
  *
- *   http://www.apache.org/licenses/LICENSE-2.0
+ *     http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
@@ -19,28 +19,26 @@
  * software solely pursuant to the terms of the relevant commercial agreement.
  */
 
-package io.crate.exceptions.scoped.table;
+package io.crate.exceptions.unscoped;
+
 
 import io.crate.exceptions.ResourceUnknownException;
-import io.crate.exceptions.TableScopeException;
-import io.crate.metadata.RelationName;
+import io.crate.exceptions.UnscopedException;
 
-import javax.annotation.Nonnull;
-import java.util.Collections;
-import java.util.Locale;
-import java.util.Objects;
+public class ColumnUnknownException extends RuntimeException implements ResourceUnknownException, UnscopedException {
 
-public class ColumnUnknownException extends RuntimeException implements ResourceUnknownException, TableScopeException {
-
-    private final RelationName relationName;
-
-    public ColumnUnknownException(String columnName, @Nonnull RelationName relationName) {
-        super(String.format(Locale.ENGLISH, "Column %s unknown", columnName));
-        this.relationName = Objects.requireNonNull(relationName);
+    /**
+     * <code>
+     * ex) select '{"x":10}'::object['y'];<br>
+     * ColumnUnknownException[The object `{x=10}` does not contain the key `y`]<br>
+     * // The column is unnamed and there is no associated table that requires permission checks. This is the only usage currently.
+     * </code>
+     */
+    public static ColumnUnknownException columnUnknownExceptionFromUnknownRelation(String message) {
+        return new ColumnUnknownException(message);
     }
 
-    @Override
-    public Iterable<RelationName> getTableIdents() {
-        return Collections.singletonList(relationName);
+    private ColumnUnknownException(String message) {
+        super(message);
     }
 }
