+++ server/src/test/java/io/crate/integrationtests/ScalarIntegrationTest.java
@@ -28,7 +28,6 @@ import org.assertj.core.api.Assertions;
 import org.elasticsearch.test.IntegTestCase;
 import org.junit.Test;
 
-import io.crate.exceptions.scoped.table.ColumnUnknownException;
 import io.crate.testing.TestingHelpers;
 import io.crate.testing.UseJdbc;
 import io.crate.types.DataTypes;
@@ -52,7 +51,7 @@ public class ScalarIntegrationTest extends IntegTestCase {
             Assertions.assertThatThrownBy(() -> sqlExecutor.exec(
                         "SELECT unnest['x'] FROM UNNEST(['{\"x\":1,\"y\":2}','{\"y\":2,\"z\":3}']::ARRAY(OBJECT))",
                         session))
-                .isExactlyInstanceOf(ColumnUnknownException.class)
+                .isExactlyInstanceOf(io.crate.exceptions.unscoped.ColumnUnknownException.class)
                 .hasMessageContaining("The object `{y=2, z=3}` does not contain the key `x`");
             // This is documenting a bug. If this fails, it is a breaking change.
             var response = sqlExecutor.exec("SELECT [unnest]['x'] FROM UNNEST(['{\"x\":1,\"y\":2}','{\"y\":2,\"z\":3}']::ARRAY(OBJECT))",
